downcast long to an int - this is a temporary checkin before we covert all the AdPod id from int to long.
[git-p4: depot-paths = "//depot/Tumri/tas/joz/": change = 9440]

+++ src/com/tumri/joz/targeting/TargetingRequestProcessor.java
@@ -153,11 +153,11 @@ public class TargetingRequestProcessor {
         }
         else if(list.size() == 1) {
             handle = list.get(0);
-            oSpec = CampaignDB.getInstance().getOSpecForAdPod(handle.getOid());
+            oSpec = CampaignDB.getInstance().getOSpecForAdPod((int)handle.getOid());
         }
         else {
             handle = selectAdPodHandle(list);
-            oSpec = CampaignDB.getInstance().getOSpecForAdPod(handle.getOid());
+            oSpec = CampaignDB.getInstance().getOSpecForAdPod((int)handle.getOid());
         }
 
         return oSpec;
